ER-291 - Deleting QualificationService and moving logic into orch
