+++ src/Employer/Employer.Web/Configuration/IoC.cs
@@ -37,7 +37,6 @@ namespace Esfa.Recruit.Employer.Web.Configuration
         {
             services.AddTransient<ITrainingProviderService, TrainingProviderService>();
             services.AddTransient<IGeocodeImageService>(_ => new GoogleMapsGeocodeImageService(configuration.GetValue<string>("GoogleMapsPrivateKey")));
-            services.AddSingleton<IQualificationsService, QualificationService>();
         }
 
         private static void RegisterProviderApiClientDep(IServiceCollection services, IConfiguration configuration)
